Move bert and xlnet utils to pretrained folder. Add test cases for bert and xlnet utils.
