+++ texar/modules/__init__.py
@@ -30,5 +30,3 @@ from texar.modules.classifiers import *
 from texar.modules.policies import *
 from texar.modules.qnets import *
 from texar.modules.memory import *
-from texar.modules.berts import *
-from texar.modules.xlnet import *
